Add a new podioIO library
