+++ src/CMakeLists.txt
@@ -80,8 +80,6 @@ SET(root_sources
   ROOTWriter.cc
   ROOTReader.cc
   ROOTLegacyReader.cc
-  Reader.cc
-  Writer.cc
 )
 if(ENABLE_RNTUPLE)
   list(APPEND root_sources
@@ -137,6 +135,23 @@ if(ENABLE_SIO)
   target_compile_definitions(podioSioIO PUBLIC PODIO_ENABLE_SIO=1)
 
   LIST(APPEND INSTALL_LIBRARIES podioSioIO podioSioIODict)
+endif()
+
+# --- IO
+set(io_sources
+  Writer.cc
+  Reader.cc
+  )
+
+PODIO_ADD_LIB_AND_DICT(podioIO "${core_headers}" "${io_sources}" io_selection.xml)
+target_link_libraries(podioIO PUBLIC podio::podio ROOT::Core ROOT::RIO ROOT::Tree)
+if(ENABLE_RNTUPLE)
+  target_link_libraries(podioIO PUBLIC ROOT::ROOTNTuple)
+  target_compile_definitions(podioIO PUBLIC PODIO_ENABLE_RNTUPLE=1)
+endif()
+if(ENABLE_SIO)
+  target_link_libraries(podioIO PUBLIC podio::podioSioIO)
+  target_compile_definitions(podioIO PUBLIC PODIO_ENABLE_SIO=1)
 endif()
 
 # --- Install everything
