+++ src/io_selection.xml
@@ -1,0 +1,4 @@
+<lcgdict>
+  <selection>
+  </selection>
+</lcgdict>
