+++ tests/sio_io/read_interface_sio.cpp
@@ -1,0 +1,11 @@
+#include "read_interface.h"
+
+int main(int, char**) {
+
+  auto readerSIO = podio::makeReader("example_frame_sio_interface.sio");
+  if (read_frames(readerSIO)) {
+    return 1;
+  }
+
+  return 0;
+}
